Draft: split upgrade from Draft.py
.
